test: take MicroVMFactory out of conftest.py
Move it to microvm.py, where it can be imported for other use
cases (scripts, for example).
